Убрал ненужный класс SetUserData, Проверку default setting сделал в main
