+++ Lesson-8/Lesson-8.csproj
@@ -65,7 +65,6 @@
       <DesignTimeSharedInput>True</DesignTimeSharedInput>
       <DependentUpon>Settings.settings</DependentUpon>
     </Compile>
-    <Compile Include="SetUserData.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
